fix: move core group definitions to model layer
Move coreGroups map from golang plugin (pkg/plugins/golang/internal/coregroups/)
to the model layer (pkg/model/resource/core_groups.go) and export a
CoreGroupDomain() lookup function instead of a raw map.
